+++ docs/user_guide/changelog.rst
@@ -22,6 +22,7 @@ New features
 - Deprecated ``runningVarianceAlwaysFloat`` in favour of ``runningStatisticsAlwaysFloat`` in ``poptorch.Options.Precision``,
   as this new option computes both the running mean and variance in FP32 when this option is set to `True`.
 - Add support for ``torch.scatter_add``
+- Add ``poptorch.Options.Precision.enableFloatingPointExceptions`` to control floating point exception behavior
 
 v2.0 (Poplar SDK 2.0)
 =====================